+++ herbrain/preprocessing/smoothing.py
