+++ herbrain/regression/decimation.py
